+++ demo/src/main/java/com/orange/ods/demo/ui/utilities/composable/Texts.kt
@@ -29,13 +29,13 @@ import com.orange.ods.compose.theme.OdsTheme
 import com.orange.ods.demo.R
 
 @Composable
-fun Title(@StringRes textRes: Int, modifier: Modifier = Modifier, withHorizontalPadding: Boolean = false) {
+fun Title(@StringRes textRes: Int, modifier: Modifier = Modifier, horizontalPadding: Boolean = false, topPadding: Boolean = false) {
     OdsTextH5(
         text = stringResource(textRes),
         modifier = modifier
             .fillMaxWidth()
-            .padding(top = dimensionResource(R.dimen.spacing_xl))
-            .padding(horizontal = if (withHorizontalPadding) dimensionResource(R.dimen.screen_horizontal_margin) else 0.dp)
+            .padding(top = if (topPadding) dimensionResource(R.dimen.spacing_xl) else 0.dp)
+            .padding(horizontal = if (horizontalPadding) dimensionResource(R.dimen.screen_horizontal_margin) else 0.dp)
     )
 }
 
@@ -43,8 +43,8 @@ fun Title(@StringRes textRes: Int, modifier: Modifier = Modifier, withHorizontal
 fun Subtitle(@StringRes textRes: Int, displaySurface: OdsDisplaySurface = OdsDisplaySurface.Default, withHorizontalPadding: Boolean = false) {
     val backgroundColor = when (displaySurface) {
         OdsDisplaySurface.Default -> Color.Unspecified
-        OdsDisplaySurface.Dark -> OdsTheme.darkThemeColors.background
-        OdsDisplaySurface.Light -> OdsTheme.lightThemeColors.background
+        OdsDisplaySurface.Dark -> OdsTheme.darkThemeColors.surface
+        OdsDisplaySurface.Light -> OdsTheme.lightThemeColors.surface
     }
     OdsTextSubtitle1(
         text = stringResource(textRes),